Bug: pvs returns a string not an array
The issue was that the output of the pvs command is a string, not an
array. The inject method was being called on this string and would cause
an exception. To fix this, the pvs result value is split on the newline
character to ensure that an array is being used for the inject
invocation.